# swagger/swagger.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:9090
...
paths:
...
  /api/v1/language:
    post:
      operationId: createLanguage
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguageCreateDto"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LanguageDto"
          description: Language record created
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Create a new language record
      tags:
      - Language
    put:
      operationId: updateLanguage
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguageUpdateDto"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LanguageDto"
          description: Language record updated
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Update an existing language record
      tags:
      - Language
  /api/v1/language/{id}:
    delete:
      operationId: deleteLanguage
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Language record deleted
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Delete a language record by ID
      tags:
      - Language
    get:
      operationId: getLanguageById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LanguageDto"
          description: Language record found
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Get a language record by ID
      tags:
      - Language
  /api/v1/link:
    post:
      description: Create a new link for a resume.
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkCreateDto"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LinkDto"
          description: Link created successfully
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Create a new Link
      tags:
      - Link
    put:
      description: Update the details of an existing link.
      operationId: update_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkUpdateDto"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LinkDto"
          description: Link updated successfully
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Update an existing Link
      tags:
      - Link
  /api/v1/link/{id}:
    delete:
      description: Delete a link by its ID.
      operationId: delete_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Link deleted successfully
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Delete Link by ID
      tags:
      - Link
    get:
      description: Retrieve a link by its ID.
      operationId: getById_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LinkDto"
          description: Link retrieved successfully
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
          description: Bad Request - Invalid argument
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
          description: Not Found - Data not found
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
          description: Conflict - Data conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
          description: Internal Server Error - Unknown error occurred
      summary: Get Link by ID
      tags:
      - Link
...
components:
  schemas:
    BadRequestErrorResponse:
      type: object
      description: Bad Request Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "400"
        message:
          type: string
          description: Description of the error
          example: Invalid input provided
...
    ConflictErrorResponse:
      type: object
      description: Conflict Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "409"
        message:
          type: string
          description: Description of the error
          example: Data Conflict
...
    InternalServerErrorResponse:
      type: object
      description: Internal Server Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "500"
        message:
          type: string
          description: Description of the error
          example: Internal Server Error
...
    LanguageCreateDto:
      type: object
      description: Language Record Create Model
      properties:
        name:
          type: string
          description: Language name
          maxLength: 255
          minLength: 0
        proficiency:
          type: string
          description: Language proficiency level
          enum:
          - A1
          - A2
          - B1
          - B2
          - C1
          - C2
        resumeId:
          type: integer
          format: int64
          description: Resume ID
      required:
      - name
    LanguageDto:
      type: object
      description: Language Record Model
      properties:
        id:
          type: integer
          format: int64
          description: Language ID
        name:
          type: string
          description: Language name
        proficiency:
          type: string
          description: Language proficiency level
          enum:
          - A1
          - A2
          - B1
          - B2
          - C1
          - C2
    LanguageUpdateDto:
      type: object
      description: Language Record Update Model
      properties:
        id:
          type: integer
          format: int64
          description: Language ID
        name:
          type: string
          description: Language name
          maxLength: 255
          minLength: 0
        proficiency:
          type: string
          description: Language proficiency level
          enum:
          - A1
          - A2
          - B1
          - B2
          - C1
          - C2
        resumeId:
          type: integer
          format: int64
          description: Resume ID
    LinkCreateDto:
      type: object
      description: Link Create Model
      properties:
        platformName:
          type: string
          description: Platform name
          maxLength: 255
          minLength: 0
        profileUrl:
          type: string
          description: Profile URL on the platform
          maxLength: 255
          minLength: 0
        resumeId:
          type: integer
          format: int64
          description: Resume ID
      required:
      - platformName
      - profileUrl
    LinkDto:
      type: object
      description: Link Model
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the link
        platformName:
          type: string
          description: Platform name
        profileUrl:
          type: string
          description: Profile URL on the platform
    LinkUpdateDto:
      type: object
      description: Link Update Model
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the link
        platformName:
          type: string
          description: Platform name
          maxLength: 255
          minLength: 0
        profileUrl:
          type: string
          description: Profile URL on the platform
          maxLength: 255
          minLength: 0
        resumeId:
          type: integer
          format: int64
          description: Resume ID to associate the link with
    NotFoundErrorResponse:
      type: object
      description: Not Found Error Response
      properties:
        code:
          type: string
          description: HTTP status code of the error
          example: "404"
        message:
          type: string
          description: Description of the error
          example: Not Found
...